+++ src/components/modal/ModalCreator.jsx
@@ -1,5 +1,5 @@
 import { createPortal } from "react-dom";
-import ReservationModal from "./Modal";
+import ReservationModal from "./ReservationModal";
 
 // eslint-disable-next-line react/prop-types
 export default function ModalCreator({ close }) {
